+++ WodiLib/WodiLib.Test/Map/ValueObject/MapEventIdTest.cs
@@ -123,7 +123,7 @@ namespace WodiLib.Test.Map
             var errorOccured = false;
             try
             {
-                castValue = (int) instance;
+                castValue =  instance;
             }
             catch (Exception ex)
             {
